__app__/requirements.txt
azure-functions
numpy == 1.17.4
requests == 2.22.0
pandas == 0.25.3
azure-storage-blob == 12.1.0
shapely == 1.6.4.post2
azure-identity == 1.2.0
azure-keyvault-secrets == 4.0.0
